src/requirements.txt
# Core ML
torch>=2.0.0
numpy

# RL environment
stable-retro
torchrl
gymnasium
torchvision

# Utilities
einops
tqdm
wandb

# For video recording
moviepy
imageio
imageio-ffmpeg
